Changes according to mentor remarks
